requirements.txt
langchain==0.1.6
langchain_community==0.0.19
langchain_core==0.1.22
matplotlib==3.8.2
nltk==3.8.1
python-dotenv==1.0.1
Requests==2.31.0
Telethon==1.34.0
